Add Swagger UI + rework order form
This updates adds the "/api" page made with Swagger to detail all the controllers that exist in the solution. Plus, it adds admin controller prototypes and reworks the shop item order form so that it doesn't redirect the user to the POST path, instead just sending the data (mail will be implemented in the future).